check point:该主分支包含了udpserver和tcpserver从这里产生新的分支，将在vs工程的filter中移除tcpserver和udpserver新的分支称为simple-branch

+++ mandis-p2p/kadbucket.cpp
@@ -17,7 +17,7 @@ namespace p2pnet {
         if (cnt >= kad_bucket_->size()) {
             for (iter = kad_bucket_->begin(); iter != kad_bucket_->end(); iter++)
                 container.push_back(*iter);
-            return kad_bucket_->size();
+            return (int)kad_bucket_->size();
         }
 
         struct DigestCompare {
